Javafiles and JavaRoots will now be calculated beforehand and then used within our Refactorong classes
